pyproject.toml
[project]
name = "howistheweather"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dash>=3.0.4",
    "google-genai>=1.19.0",
    "pytest>=8.4.0",
    "python-dotenv>=1.1.0",
    "requests-mock>=1.12.1",
]
